Add warning about the documentation move
Based on the read the docs documentation there doesn't seem to be a way
to automatically redirect from retworkx.readthedocs.io to
qiskit.org/documentation. The best option I could come up with is to add
a warning to the documentation hosted on readthedocs. This commit adds a
warning to the top of every page saying the documentation on rtd is out
of date and users should go to the new home for the documentation.